Refactor WorkerManager to avoid circular import chain for headless workflows
